Rename beta model files
